Added "climb.nss" script, though it's really a "move somehow" script. Special/Custom 4 now has a Seat and Climb object.
Climb object name can be renamed or changed description to give better idea of what it does.
It might be "Climb Up DC 10", or "Cross Log DC 5", or just "Climb across ledge DC 0". "Climbing..." overhead text is changed to "Attempting...".
LocalInt on object for DC. Maybe LocalString for action word.
Also be sure to rename the Tag if two sets will be next to eachother, the script simply looks for the nearest object with the same Tag, so these work in sets automatically.