Beholder and formians, plus misc fixes
Added a new task and bounty given by the Seer in Lith My'athar. You might want to be quite high level (11 or 12) for this one, though.
Increased beholder bite damage from 2d4 to 2d6.
Increased beholder Strength from 11 to 14.
If you are below level 11, the Seer now comments that trying to stop the Valsharess might be harder until you are more experienced. She doesn't stop you trying, though.
Added dialogue to the Seer to turn in the Hive Mother quest.
Mischa now uses Remove Disease on nearby diseased friendlies when out of combat. She's a very helpful soul and will do her best to cure all those in need.
Added an escape lever to Layenne's Tomb so it is no longer possible to get stuck if someone else opened the locked door and you logged out inside.
Fixed Ferdinand's two tasks not appearing in the quest log after logging out and back in.
Fix towards ferrymasters scamming money and not putting passengers on their boat.
Adjusted the Mummy of the Damned - it's now behind a door at the other end of the Warrens. The old door still offers a faster way out once it's dead, though.