Minor fixes
Fix towards Bim acting very strangely when a player completes the academy sequence while he is already in a party with someone else.
Being in a state where your character cannot be saved should now prompt a message every time you enter an area.
Reduced the prevalence of mummies in the encounter set used in and around the Valley of the Dead.
Usage of advantageous developer functions now outputs webhooks in the name of transparency.
Reduced the number of spawns in the Valley of the Dead warrior tomb.
Increased the odds of random creatures getting Weapon Finesse if it is applicable to them.
Pickpocketing attempts are now logged due to potential involvement in the busy NPC difficulties.
Ranger adventurers should now come with a set of random favored enemies.
Items in henchman stores now transfer to the henchman's local pawnshop when dismissed, but not if they are killed. Phantom stores of old henchmen also no longer persist until server restart.
Suppressed some logging related to random spellbooks.
Logging in now notifies you of your current respawn location. Characters that haven't ever asked a priest to set their location (and end up in Neverwinter) get no message.
Reduced the number of goblins that spawn off Neverwinter's East Road.
Reduced the number of goblins that spawn in the first section of the Goblin Cave.